% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{GPT 질문 함수 (완전체 버전)}
\usage{
ask(question, history = 1, data = 1)
}
\arguments{
\item{question}{질문 문자열}

\item{history}{콘솔 기록 첨부 여부 (0/1)}

\item{data}{Environment 개괄정보 첨부 여부 (0/1)}
}
\description{
GPT 질문 함수 (완전체 버전)
}
