% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R, R/key_management.R
\name{encrypt_api_key}
\alias{encrypt_api_key}
\title{API 키 암호화 함수}
\usage{
encrypt_api_key(key, password)

encrypt_api_key(key, password)
}
\arguments{
\item{key}{API 키 문자열}

\item{password}{비밀번호 문자열}
}
\description{
API 키 암호화 함수

API 키 암호화 함수
}
