% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{gpt}
\alias{gpt}
\title{GPT 질문 함수}
\usage{
gpt(question, history = 1, data = 1)
}
\arguments{
\item{question}{질문 문자열}

\item{history}{콘솔 기록 첨부 여부 (0/1)}

\item{data}{데이터프레임 첨부 여부 (0/1)}
}
\description{
GPT 질문 함수
}
