% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heip.R
\name{heip}
\alias{heip}
\title{GPT 질문 함수 (이제 heip 이름으로)}
\usage{
heip(question, history = 1, data = 1)
}
\arguments{
\item{question}{질문 문자열}

\item{history}{콘솔 기록 첨부 여부 (0/1)}

\item{data}{데이터프레임 첨부 여부 (0/1)}
}
\description{
GPT 질문 함수 (이제 heip 이름으로)
}
