% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr.R
\name{smr}
\alias{smr}
\title{요약 출력 없이 GPT에 데이터 요약과 질문만 전송하는 smr()}
\usage{
smr(data, max_lines = 50)
}
\arguments{
\item{data}{c("x", "y") 형식으로 데이터 객체 이름 전달}

\item{max_lines}{각 블록별 최대 출력 줄 수 제한}
}
\description{
요약 출력 없이 GPT에 데이터 요약과 질문만 전송하는 smr()
}
